\name{DDIwR-internal}

\alias{checkType}
\alias{cleanup}
\alias{convertibble}
\alias{getDecimals}
\alias{getDelimiter}
\alias{getEnter}
\alias{getFiles}
\alias{getFormat}
\alias{getXML}
\alias{splitrows}
\alias{test}
\alias{tibbleMetadata}
\alias{treatPath}
\alias{writeRlist}

\title{Internal Functions}

\description{
  The above functions are internal in the DwB package which are not designed to 
  be called directly by the user.
}

\keyword{internal}
