% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCategories.R
\name{makeCategories}
\alias{makeCategories}
\title{Create the \code{catgry} elements for a particular variable}
\usage{
makeCategories(metadata)
}
\arguments{
\item{metadata}{A list of two or three components: \code{labels},
\code{na_values} and/or \code{na_range}}
}
\value{
A standard R list of class \verb{"DDI}.
}
\description{
Utility function to create the \code{catgry} elements, as well as all
necessary sub-elements (e.g. \code{catValu}, \code{labl}, \code{varFormat}) along with their
associated XML attributes.
}
\author{
Adrian Dusa
}
