% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNotes.R
\name{makeDataNotes}
\alias{makeDataNotes}
\title{Create a \code{notes} element for the dataset.}
\usage{
makeDataNotes(data)
}
\arguments{
\item{data}{An R dataframe.}
}
\value{
A standard \code{notes} DDI element.
}
\description{
Create the \code{notes} element to embed a serialized, gzip-ed version of the data
in the \code{fileDscr} section of the \code{codeBook}.
}
\author{
Adrian Dusa
}
