% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFor.R
\name{searchFor}
\alias{searchFor}
\title{Search for key words}
\usage{
searchFor(
  word,
  where = c("everywhere", "description", "examples", "attributes")
)
}
\arguments{
\item{word}{Character, either a word or a regular expression.}

\item{where}{Character, in which sections to search for.}
}
\value{
Character vector of DDI element names.
}
\description{
Search function to return elements that contain a certain
word of regular expression pattern.
}
\author{
Adrian Dusa
}
