% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDetails.R
\name{showDetails}
\alias{showDetails}
\alias{showDescription}
\alias{showAttributes}
\alias{globalAttributes}
\alias{showExamples}
\alias{showRelations}
\alias{showLineages}
\title{Describe what a DDI element is}
\usage{
showDetails(x, ...)

showDescription(x, ...)

showAttributes(x, name = NULL, ...)

globalAttributes()

showExamples(x, ...)

showRelations(x, ...)

showLineages(x, ...)
}
\arguments{
\item{x}{Character, a DDI Codebook element name.}

\item{...}{Other arguments, mainly for internal use.}

\item{name}{Character, print only a specific element (name)}
}
\description{
Describe what a DDI element is
}
\details{
All arguments having predefined values such as "(Y | N) : N" are
mandatory if the element is used
}
\examples{
showDetails("codeBook")

showAttributes("catgry")

showExamples("abstract")

showLineages("titl")

}
\author{
Adrian Dusa
}
