\name{transfer}

\alias{transfer}

\title{Transfer / convert between software packages}

\description{
Something similar to StatTransfer
}

\usage{
transfer(from, to, embed = FALSE, execute = "", ...)
}

\arguments{
  \item{from}{A path to a file, or a tibble object}
  \item{to}{A path to a file or the name of a software package, see the Details section}
  \item{embed}{Boolean, embed the data when generating a DDI XML file}
  \item{execute}{Path to the executable file to run the batch script}
  \item{...}{Additional parameters passed to exporting functions, see the Details section}
}

\details{
Some more details here.
}



\keyword{functions}
