% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCodebook.R
\name{updateCodebook}
\alias{updateCodebook}
\title{Update Codebook.}
\usage{
updateCodebook(xmlfile, codeBook, ...)
}
\arguments{
\item{xmlfile}{A path to a DDI Codebook XML document.}

\item{codeBook}{An R object containing a \code{codeBook} element.}
}
\description{
Update an XML file containing a DDI Codebook.
}
\details{
This function replaces entire Codebook sections. Any such section present in
the R object will replace the corresponding section from the XML document.
}
\author{
Adrian Dusa
}
