% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMList.R
\name{XMList}
\alias{XMList}
\alias{fromXMList}
\alias{toXMList}
\title{Convert to and from an XML list of class \code{"xml_document"}, as used by package
\strong{xml2}.}
\usage{
fromXMList(xmlist, ...)

toXMList(element)
}
\arguments{
\item{xmlist}{An R object of class \code{"xml_document"}}

\item{...}{Other arguments, mainly for internal use.}

\item{element}{A standard DDI Codebook list element.}
}
\value{
\code{fromXMList()} returns a standard list element of class \code{"DDI"}, and
\code{toXMList()} converts back to an XML list to be written into an ".xml" file.
}
\description{
to add
}
\details{
The argument \code{xmlist} received an object created by the function
\code{read_xml()} in package \strong{xml2}. Both functions are useful to import /
export from a DDI Codebook expressed in XML, to a standard R list of class
\code{"DDI"}.
}
\author{
Adrian Dusa
}
