\name{DDIwR-package}

\alias{DDIwR-package}

\alias{DDIwR}

\docType{package}

\title{
Useful functions for various DDI (Data Documentation Initiative) related outputs.
}

\description{
This package provides various functions to read DDI based metadata documentation,
and write dedicated setup files for R, SPSS, Stata and SAS to read an associated
.csv file containing the raw data, apply labels for variables and values and
also deal with the treatment of missing values.

The research leading to these results has received funding from the European Union's
Seventh Framework Programme (FP7/2007-2013) under grant agreement no. 262608
(DwB - Data without Boundaries)
}

\details{
    \tabular{ll}{
    Package: \tab DDIwR\cr
    Type:    \tab Package\cr
    Version: \tab 0.2-0\cr
    Date:    \tab 2014-10-02\cr
    License: \tab GPL (>= 2)\cr
    }
}

\author{

\bold{Authors}:\cr
Adrian Dusa\cr
Department of Sociology\cr
University of Bucharest\cr
\email{dusa.adrian@unibuc.ro}

\bold{Maintainer}:\cr
Adrian Dusa
}

\keyword{package}

