\name{DDIwR internal functions}

\alias{checkType}
\alias{cleanup}
\alias{collectMetadata}
\alias{getDecimals}
\alias{getDelimiter}
\alias{getEnter}
\alias{getFiles}
\alias{getFormat}
\alias{getXML}
\alias{make_labelled}
\alias{replaceTicks}
\alias{splitrows}
\alias{test}
\alias{treatPath}
\alias{writeMetadata}
\alias{writeRlist}

\title{Internal Functions}

\description{
  The above functions are internal in the DwB package which are not designed to 
  be called directly by the user.
}

\keyword{internal}
