% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Extract metadata information}
\usage{
getMetadata(x, encoding = "UTF-8", ...)
}
\arguments{
\item{x}{A path to a file, or a data frame object}

\item{encoding}{The character encoding used to read a file}

\item{...}{Additional arguments for this function (internal use only)}
}
\value{
An R list roughly equivalent to a DDI Codebook, containing all variables,
their corresponding variable labels and value labels, and (if applicable)
missing values if imported and found.
}
\description{
Extract a list containing the variable labels, value labels and any
available information about missing values.
}
\details{
This function reads an XML file containing a DDI codebook version 2.6, or an
SPSS or Stata file and returns a list containing the variable labels, value
labels, plus some other useful information.

If the input is a dataset, it will extract the variable level metadata
(labels, missing values etc.). From a DDI XML file, it will import all
metadata elements, the most expensive being the data description. This
element can be skipped by deactivating the \code{dataDscr} argument.

It additionally attempts to automatically detect a type for each variable:
\tabular{rl}{
\strong{\code{cat}}: \tab categorical variable using numeric values\cr
\strong{\code{catchar}}: \tab categorical variable using character values\cr
\strong{\code{catnum}}: \tab categorical variable for which numerical summaries\cr
\tab can be calculated (ex. a 0...10 Likert response scale)\cr
\strong{\code{num}}: \tab numerical\cr
\strong{\code{numcat}}: \tab numerical variable with few enough values (ex. number of
children)\cr
\tab for which a table of frequencies is possible in addition to
frequencies
}

For the moment, only DDI Codebook is supported, but DDI Lifecycle is planned
to be implemented.
}
\examples{
x <- data.frame(
    A = declared(
        c(1:5, -92),
        labels = c(Good = 1, Bad = 5, NR = -92),
        na_values = -92
    ),
    C = declared(
        c(1, -91, 3:5, -92),
        labels = c(DK = -91, NR = -92),
        na_values = c(-91, -92)
    )
)

getMetadata(x)

}
\author{
Adrian Dusa
}
