% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDDITree.R
\name{makeDDITree}
\alias{makeDDITree}
\title{Build the DDI Codebook tree structure}
\usage{
makeDDITree(root = "codeBook")
}
\arguments{
\item{root}{Character, the root element to start from (default \code{"codeBook"}).}
}
\value{
A nested list representing the element tree.
}
\description{
Builds a tree-shaped representation of the DDI Codebook 2.6 element
hierarchy using the in-package structure definition.
}
\examples{

# Build the tree as an R list
tree <- makeDDITree()

}
