% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSchema.R
\name{updateSchema}
\alias{updateSchema}
\title{Updates the internal DDI Codebook schema object.}
\usage{
updateSchema(xsd = NULL, return = FALSE)
}
\arguments{
\item{xsd}{A path to the Codebook XML Schema file.}

\item{return}{Return an R object representing the schema instead of updating the internal one.}
}
\description{
Rebuilds the internal schema object, from a (newer) XML Schema codebook.xsd file.
}
\details{
Releasing a new stable version of the DDI Codebook takes about 10 years. There are
numerous elements and attributes that have to work together, and most importantly the Codebook
has to obey the backward compatibility rule. Until a new version is released, the Codebook
schema is incrementally modified by the DDI Alliance, in their GitHub repository.

This function is intended to update the internal DDI Codebook schema object, parsing the
latest version of the Codebook XML Schema file.

Unless the codebook.xsd file is provided by the user, the function will attempt to read it
from the DDI Alliance GitHub repository, located at:
\url{https://github.com/ddialliance/ddi-c_2}
}
\author{
Adrian Dusa
}
